🧬 Migrate the data model about entire configuration back to *__init__* module.
