+++ pymock_api/__init__.py
@@ -3,4 +3,4 @@
 content ...
 """
 
-from .model.api_config.api_config import APIConfig
+from .model.api_config import APIConfig
