+++ test/unit_test/model/swagger_config.py
@@ -8,7 +8,8 @@ from typing import List, Tuple
 
 import pytest
 
-from pymock_api.model import APIConfig, MockAPI
+from pymock_api import APIConfig
+from pymock_api.model import MockAPI
 from pymock_api.model.api_config import _Config
 from pymock_api.model.api_config.apis import APIParameter as PyMockAPIParameter
 from pymock_api.model.enums import ResponseStrategy
